// backend/src/modules/credit-note/enums/credit-note-status.enum.ts
export enum CREDIT_NOTE_STATUS {
  Nonexistent = 'creditNote.status.non_existent',
  Draft = 'creditNote.status.draft',
  Sent = 'creditNote.status.sent',
  Validated = 'creditNote.status.validated',
  Paid = 'creditNote.status.paid',
  PartiallyPaid = 'creditNote.status.partially_paid',
  Unpaid = 'creditNote.status.unpaid',
  Expired = 'creditNote.status.expired',
}